+++ src/Compilers/Lua/Portable/Errors/ErrorCode.cs
@@ -50,6 +50,7 @@
         // Using part instead of term here because it's more user friendly.
         ERR_InvalidExpressionPart = 1011,
         ERR_InvalidStatement = 1012,
+        ERR_CompoundAssignmentNotSupportedInLuaVersion = 1013,
 
         // MessageProvider stuff
         ERR_BadDocumentationMode = 2000,