+++ src/Compilers/Lua/Portable/LuaResources.Designer.cs
@@ -133,6 +133,15 @@ namespace Loretta.CodeAnalysis.Lua {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Compound assignments are not supported in this lua version.
+        /// </summary>
+        internal static string ERR_CompoundAssignmentNotSupportedInLuaVersion {
+            get {
+                return ResourceManager.GetString("ERR_CompoundAssignmentNotSupportedInLuaVersion", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Constant represents a value either too large or too small for a double precision floating-point number.
         /// </summary>
         internal static string ERR_DoubleOverflow {